% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aavenger_sites_random_match.R
\name{aavenger_sites_random_match}
\alias{aavenger_sites_random_match}
\title{AAvengeR Sites Random Match}
\usage{
aavenger_sites_random_match(
  aavenger_sites,
  chromosome_lengths,
  random_seed_value,
  match_row_number_modifier = 3
)
}
\arguments{
\item{chromosome_lengths}{A \code{data.frame} object. See \code{fasta_to_chromosome_lengths()}.}

\item{random_seed_value}{A \code{numeric} value to use as a seed for the random number generator function \code{r.seed()}.}

\item{match_row_number_modifier}{A \code{integer} value that will be used to multiply the number of rows in the insertion \code{heatmap_group} to obtain the number of sites to sample for the random match.}

\item{matched_aavenger_sites_df}{A \code{data.frame} object containing specifically formatted match and insertion sites for each heatmap_group. See \code{format_aavenger_sites()}.}
}
\value{
A \code{data.frame} object of random-matches for each unique \code{heatmap_group} in the supplied \code{aavenger_sites}. The columns will be match the output of \code{format_aavenger_sites()}.
}
\description{
Create a random-matched dataframe for each \code{heatmap_group} combination in \code{aavenger_sites}
}
\details{
This function acts as a wrapper for \code{sample_genome()}.
}
\examples{
aavenger_sites_random_match(aavenger_sites, chromosome_lengths, random_seed_value)

}
