% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_integrations_in_feature_window.R
\name{count_integrations_in_feature_window}
\alias{count_integrations_in_feature_window}
\title{Count integrations in a genomic feature window}
\usage{
count_integrations_in_feature_window(
  df_sites,
  feature_file,
  overlap_window_size
)
}
\arguments{
\item{df_sites}{A \code{data.frame} object containing genomic site information. Can come from two sources:
Generated by \code{format_aavenger_sites()} or \code{sample_genome()}.}

\item{feature_file}{Path to either an \code{RDS} file or the path to an \code{RData} file containing genomic feature information.}

\item{overlap_window_size}{The size of the window around each site where integrations will be counted.}
}
\value{
A numeric vector containing the count of integrations, total integration sites, and the percentage of integration sites.
}
\description{
This function counts the number of genomic integration sites within a specified feature window.
}
\examples{
result <- count_integrations_in_feature_window(test_df, test_feature_file, test_overlap_window_size)

}
\seealso{
\code{\link{readRDS}}, \code{\link{GenomicRanges::makeGRangesFromDataFrame}}, \code{\link{GenomicRanges::countOverlaps}}
}
