% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_repeat_class_per_cluster.R
\name{count_repeat_class_per_cluster}
\alias{count_repeat_class_per_cluster}
\title{Count Repeat Classes Per Cluster}
\usage{
count_repeat_class_per_cluster(
  aavenger_mhc_repeat_overlaps,
  repeat_column_name = "repeat_class_subject"
)
}
\arguments{
\item{aavenger_mhc_repeat_overlaps}{A data frame containing AAvengeR MHC repeat overlaps, with columns for repeat classifications and query data.}

\item{repeat_column_name}{A character string specifying the name of the column that contains repeat classification information. Default is \code{"repeat_class_subject"}.}
}
\value{
A data frame with counts and percentages of each repeat class per cluster, including:
\describe{
\item{trial}{Trial identifier.}
\item{subject}{Subject identifier.}
\item{sample}{Sample identifier.}
\item{clusterID}{Cluster identifier.}
\item{repeat_class}{Simplified repeat class.}
\item{repeat_class_count}{Count of each repeat class within the cluster.}
\item{total_repeat_detections}{Total repeat detections in the cluster.}
\item{percentage_of_repeat_detections}{Percentage of each repeat class in the cluster.}
}
}
\description{
This function calculates the count and percentage of different repeat classes within each cluster in the AAvengeR MHC data. It aggregates data by cluster and computes the proportion of each repeat class per cluster.
}
\details{
Counts the number of repeats found in a multihit clusters object outputted by AAVengeR.
}
