% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_aavenger_table.R
\docType{data}
\name{example_aavenger_table}
\alias{example_aavenger_table}
\title{Example AAVengeR table output}
\format{
A data frame with 2000 rows and 15 columns:
\describe{
\item{trial}{trial name is testset}
\item{subject}{subject id 1 or 2}
\item{sample}{sampling repeats}
\item{refGenome}{reference genome used}
\item{posid}{AAVengeR position ID}
\item{UMIs}{unique molecular identifiers}
\item{sonicLengths}{unique sonic breaks}
\item{nRepsObs}{Number of observed repetitions}
\item{flags}{flags used}
\item{vectorFastaFile}{vector used for gene delivery}
\item{rep1-UMIs}{redundant}
\item{rep1-sonicLengths}{redundant}
\item{rep1-reads}{redundant}
\item{rep1-repLeaderSeq}{redundant}
\item{maxLeaderSeqDist}{redundant}
}
}
\usage{
example_aavenger_table
}
\description{
A table made to mimic AAVengeR's tabular sites output.
}
\examples{
  example_aavenger_table
}
\keyword{datasets}
