% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_fasta_chromosome.R
\docType{data}
\name{example_fasta_chromosome}
\alias{example_fasta_chromosome}
\title{Example fasta file with ACTGN}
\format{
DNAStringSet object of length 2:
\describe{
\item{width}{Length of sequence)}
\item{seq}{sequence}
}
}
\usage{
example_fasta_chromosome
}
\description{
A Biostrings::DNAStringSet object with two chromosomes of 1000 bp length and
differing GC percentages (50 and 60 percent, respectively).
}
\examples{
  example_fasta_chromosome
}
\keyword{datasets}
