% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_H3K79me2_rdata.R
\docType{data}
\name{example_H3K79me2_rdata}
\alias{example_H3K79me2_rdata}
\title{RData features table}
\format{
A GRanges object with 1843018 ranges, 3 columns and 1 metadata column:
\describe{
\item{seqnames}{sequence name e.g. chr1 through chrX)}
\item{ranges}{basepair ranges of feature}
\item{strand}{strandedness of features. default *}
\item{mid}{middle range of ranges}
}
}
\usage{
example_H3K79me2_rdata
}
\description{
A Genomic Ranges feature table object stored as RData file type. Named epigenData
}
\examples{
  example_H3K79me2_rdata
}
\keyword{datasets}
