% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_hg38_chromosome_lengths.R
\docType{data}
\name{example_hg38_chromosome_lengths}
\alias{example_hg38_chromosome_lengths}
\title{Dataset of hg38 chromosome lengths}
\format{
A data frame with 24 rows and 2 columns:
\describe{
\item{seqname}{chromosome name}
\item{chr_length}{length of chromosome}
}
}
\usage{
example_hg38_chromosome_lengths
}
\description{
Chromosome lengths for all 22 numbered chromosomes + X and Y chromosomes.
}
\examples{
  example_hg38_chromosome_lengths
}
\keyword{datasets}
