% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_refGene_rds.R
\docType{data}
\name{example_refGene_rds}
\alias{example_refGene_rds}
\title{Rds features table}
\format{
A GRanges object with 44380 ranges, 3 columns and 0 metadata column:
\describe{
\item{seqnames}{sequence name e.g. chr1 through chrX)}
\item{ranges}{basepair ranges of feature}
\item{strand}{strandedness of features. default *}
}
}
\usage{
example_refGene_rds
}
\description{
A Genomic Ranges feature table object stored as rds file type.
}
\examples{
  example_refGene_rds
}
\keyword{datasets}
