% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_refseq_full_table.R
\docType{data}
\name{example_refseq_full_table}
\alias{example_refseq_full_table}
\title{RefSeq gene table}
\format{
A GRanges object with 1000 rows and 16 columns:
\describe{
\item{bin}{internal indexing}
\item{name}{Name of gene}
\item{chrom}{Reference chromosome}
\item{strand}{Genomic strand of entry}
\item{txStart}{Start of transcription (end position for minus strand)}
\item{txEnd}{End of transcription (start position for minus strand)}
\item{cdsStart}{Start of coding region (end position for minus strand)}
\item{cdsEnd}{End of coding region (start position for minus strand)}
\item{exonCount}{Number of exons}
\item{exonStarts}{Start of exond (end position for minus strand)}
\item{exonEnds}{End of coding region (start position for minus strand)}
\item{score}{score}
\item{name2}{Alternate name}
\item{cdsStartStat}{Status of CDs start annotation (none, unknown, incomplete, complete)}
\item{cdsEndStat}{Status of CDs end annotation (none, unknown, incomplete, complete)}
\item{exonFrames}{Reading frame of the CDS region of the exon, in the direction of transcription (0, 1, 2) or -1 if no CDS region.}
}
}
\usage{
example_refseq_full_table
}
\description{
A Refseq data table containing coordinates for genes
}
\details{
Based on https://hgdownload.soe.ucsc.edu/goldenPath/hg38/database/ncbiRefSeq.txt.gz
}
\examples{
  example_refseq_full_table
}
\keyword{datasets}
