% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental_counts_in_features.R
\name{experimental_counts_in_features}
\alias{experimental_counts_in_features}
\title{Calculate experimental integration counts in genomic features for multiple scenarios}
\usage{
experimental_counts_in_features(
  df_sites,
  chromosome_lengths,
  overlap_window_sizes,
  feature_files_to_process,
  number_of_cores = parallel::detectCores()
)
}
\arguments{
\item{df_sites}{A \code{data.frame} containing experimental genomic site information, including columns \code{seqname}, \code{start}, \code{end}, \code{width}, \code{strand}, \code{mid}, \code{type}, \code{heatmap_group}.
#'}

\item{overlap_window_sizes}{A numeric vector containing different sizes of overlap windows around each integration site.}

\item{feature_files_to_process}{A character vector containing absolute paths to feature files for processing. Must be \code{GenomicRanges} objects in .rds or RData (named epigenData if RData) format. See \code{example_refGene_rds}}

\item{number_of_cores}{The number of CPU cores to use for parallel processing. Defaults to the number of available cores.}
}
\value{
A data frame containing experimental integration counts and percentages in features, including overlap window size, feature track file, heatmap group,
count of integrations in features, total integration sites, and percentage of integrations in features.
}
\description{
This function calculates the number of experimental genomic integrations within specified genomic features for multiple scenarios defined by varying parameters.
}
\examples{
# Example usage:
df_sites <- format_aavenger_sites(unformatted_aavenger_sites)
overlap_window_sizes <- c(10, 20, 30)
feature_files <- c("path/to/feature_file1.rds", "path/to/feature_file2.rds")
experimental_counts_in_features(df_sites, chromosome_lengths, overlap_window_sizes, feature_files)

}
