% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_raw_refseq_table.R
\name{extend_raw_refseq_table}
\alias{extend_raw_refseq_table}
\title{Extend raw RefSeq table}
\usage{
extend_raw_refseq_table(refseq_table, number_of_cores = 8)
}
\arguments{
\item{refseq_table}{A data frame containing raw RefSeq data.}

\item{number_of_cores}{The number of CPU cores to be used for parallel processing. Default is 8.}
}
\value{
A data frame containing extended RefSeq data with columns start, end, feature, and name
}
\description{
This function extends a raw RefSeq table by splitting it into chunks based on the number of cores specified,
processing each chunk in parallel, and then combining the results into a single table.
}
\details{
Calls on InvestigateIntegrations::separate_gene_features()
Dataframe must have the following columns:
strand genomic strand
chrom chromosome name
exonStarts comma delimited list of exon positions and end with a comma (character)
exonEnds comma delimited list of exon positions and end with a comma (character)
exonCount number of exons in exonStarts/exonEnds (numeric)
cdsStartStat either cmpl or none (character)
txStart transcription start (numeric)
txEnd transcription end (numeric)
cdsStart coding start (numeric)
cdsEnd coding end (numeric)
name identifier

All numeric positions are being read in 5'->3'
Returns a dataframe with each row being a range for the category of UTR_left, UTR_right, exon, and intron
}
