% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta_to_chromosome_lengths.R
\name{fasta_to_chromosome_lengths}
\alias{fasta_to_chromosome_lengths}
\title{Fasta to Chromosome Lengths}
\usage{
fasta_to_chromosome_lengths(genome_fasta)
}
\arguments{
\item{genome_fasta}{A \code{ShortRead} object with \verb{@id} and \verb{@width} values.}
}
\value{
A \code{data.frame} object with two columns: \code{seqname} and \code{chr_length}.
}
\description{
Converts a fasta file into a dataframe with two columns: chromosome name and chromosome length.
}
\details{
This function explicitly shows how a genomic fasta is converted into the chromosome_lengths input used in other functions. A genomic fasta
file read in with \code{ShortRead::readFasta()} will generate the required inputs for this function.
}
\examples{
fasta_to_chromosome_lengths(genome_fasta)

}
