% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_groups_of_integration_positions.R
\name{find_groups_of_integration_positions}
\alias{find_groups_of_integration_positions}
\title{Find Groups of Integration Positions}
\usage{
find_groups_of_integration_positions(
  positions_to_check,
  overlap_search_range = 10000,
  min_group_size = 1
)
}
\arguments{
\item{positions_to_check}{A numeric vector of positions to be grouped. Must contain more than one value.}

\item{overlap_search_range}{A numeric value indicating the maximum allowable distance between
consecutive positions for them to be considered part of the same group. Default is 10,000.}

\item{min_group_size}{An integer specifying the minimum size of a group to include in the results.
Groups smaller than this size will be excluded. Default is 1 (no filtering).}
}
\value{
A list of numeric vectors, where each vector represents a group of positions.
If \code{min_group_size} is greater than 1, only groups meeting this size requirement will be included.
}
\description{
This function identifies groups of positions where consecutive values in a sorted vector
are within a specified overlap range. It can also filter groups based on a minimum group size.
}
