% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_aavenger_sites.R
\name{format_aavenger_sites}
\alias{format_aavenger_sites}
\title{Format AAVengeR sites}
\usage{
format_aavenger_sites(aavenger_sites_df)
}
\arguments{
\item{aavenger_sites_df}{A \code{data.frame} object. Must at least contain the following columns:
\describe{
\item{posid}{The integration site position in the AAVengeR format.}.
\item{trial}{The trial name.}
\item{subject}{The subject name.}
\item{sample}{The sample name.}}}
}
\value{
A \code{data.frame} object with the following columns: \code{seqname}, \code{start}, \code{end}, \code{width}, \code{strand}, \code{mid}, \code{type}, \code{heatmap_group}
}
\description{
Formats an AAvengeR sites \code{dataframe} object so that it is usable in all downstream functions.
}
\details{
This function will paste together the trial, subject, and sample values into a single heatmap_group value. It will also assign
\code{type} to be \code{insertion}.
}
\examples{
format_aavenger_sites(aavenger_sites_df)

}
