% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_hot_roc_result.R
\name{format_hot_roc_result}
\alias{format_hot_roc_result}
\title{Format hotROCs Result}
\usage{
format_hot_roc_result(hot_roc_result)
}
\arguments{
\item{hot_roc_result}{The multidata object produced by \code{hotROCs::ROC.ORC()}.}
}
\value{
A \code{data.frame} object with the \code{feature_window}, the \code{heatmap_group} it belongs to, the calculated \code{p_value}, the \code{sig_p_value}, and the \code{ROC_value}
\describe{
\item{feature_window}{The genomic feature sampled and window size that was used.}
\item{heatmap_group}{The grouping used.}
\item{p_value}{The p-value calculated by \code{hotROCs::ROC.ORC()}.}
\item{sig_p_value}{Star representation of p-value.}
\item{ROC_value}{The ROC value calculated by \code{hotROCs::ROC.ORC()}.}}
}
\description{
Reformats the ROC output of \code{hotroc_compare_insertion_to_match()} into a \code{data.frame()} object.
}
\examples{
format_hot_roc_result(hot_roc_result)

}
