% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GC_percentage_per_insertion.R
\name{GC_percentage_per_insertion}
\alias{GC_percentage_per_insertion}
\title{Calculate GC Content Around Insertion Sites}
\usage{
GC_percentage_per_insertion(
  df_combined,
  genomic_sequence,
  window_sizes_to_test,
  chromosome_lengths,
  fasta_name = "genome"
)
}
\arguments{
\item{df_combined}{A dataframe containing genomic insertion sites with at least the following columns:
\describe{
\item{seqname}{Chromosome name or sequence identifier}
\item{start}{Start position of the insertion}
\item{end}{End position of the insertion}
}}

\item{genomic_sequence}{A \code{BSgenome} object containing the genomic sequence data.}

\item{window_sizes_to_test}{A numeric vector of window sizes (in base pairs) to test around each insertion site.}

\item{chromosome_lengths}{A dataframe with chromosome lengths, containing at least the following columns:
\describe{
\item{seqname}{Chromosome name or sequence identifier}
\item{chr_length}{Length of the chromosome}
}}

\item{fasta_name}{A string representing the base name for the output fasta file, default is 'genome'.}
}
\value{
A dataframe identical to \code{df_combined} but with additional columns for each window size tested.
Each new column represents the GC content calculated for that specific window size around the insertion sites.
}
\description{
This function calculates the GC content for various window sizes around genomic insertion sites.
It takes a dataframe of insertion sites, a genomic sequence, and a set of window sizes, and
returns the original dataframe augmented with columns representing the GC content for each window size.
}
