% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotroc_compare_insertion_to_match.R
\name{hotroc_compare_insertion_to_match}
\alias{hotroc_compare_insertion_to_match}
\title{hotROCs Compare Insertion to Match}
\usage{
hotroc_compare_insertion_to_match(matched_overlap_df)
}
\arguments{
\item{matched_overlap_df}{A \code{data.frame} object containing specifically formatted match and insertion sites for each heatmap_group. See output of \code{test_for_overlaps()}.}
}
\value{
Outputs the multidata object produced by \code{hotROCs::ROC.ORC()}.
}
\description{
Supplies a random-matched set of integration sites to \code{hotROCs::ROC.ORC()}.
}
\details{
This function wraps \code{hotROCs::ROC.ORC()}. It is meant to abstract away the input format details for this function as they can be confusing. The input for this function will have the same columns as the output of
\code{test_for_overlaps()} but it must include both sites of \code{type} 'match' and 'insertion' for each unique \code{heatmap_group}.
}
\examples{
hotroc_compare_insertion_to_match(matched_overlap_df)

}
