% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_percentage_in_feature_p_value.R
\name{integration_percentage_in_feature_p_value}
\alias{integration_percentage_in_feature_p_value}
\title{Calculate p-values for observed integration percentages in genomic features}
\usage{
integration_percentage_in_feature_p_value(
  simulation_results,
  experimental_results
)
}
\arguments{
\item{simulation_results}{A \code{data.frame} containing simulation results, including overlap window size, feature track file,
and simulated integration percentages. Generated by \code{rnorm_simulate_many_counts_in_features()} or \code{simulate_many_counts_in_features()}.}

\item{experimental_results}{A \code{data.frame} containing experimental integration results, including overlap window size, feature track file,
heatmap group, and observed integration percentages. Generated by \code{experimental_counts_in_features()}.}
}
\value{
A \code{data.frame} containing observed integration percentages, corresponding probabilities from simulated distributions,
and two-sided p-values for each heatmap group, feature track, and overlap window size combination.
}
\description{
This function calculates p-values for observed integration percentages in genomic features based on simulated distributions.
}
\examples{
# Example usage:
integration_percentage_in_feature_p_value(simulation_results, experimental_results)

}
