% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multihit_cluster_to_representative_repeat.R
\name{multihit_cluster_to_representative_repeat}
\alias{multihit_cluster_to_representative_repeat}
\title{Assign Representative Repeat Class to Multihit Clusters}
\usage{
multihit_cluster_to_representative_repeat(
  aavenger_mhc,
  repeat_table_grange,
  repeat_column_name = "repeat_class_subject"
)
}
\arguments{
\item{aavenger_mhc}{A data frame representing the AAvengeR MHC data, with positional and cluster information.}

\item{repeat_table_grange}{A \code{GRanges} object containing repeat regions to check for overlaps.}

\item{repeat_column_name}{A character string indicating the column name that stores repeat classification information in the overlap results. Default is \code{"repeat_class_subject"}.}
}
\value{
A data frame with the most abundant repeat class for each multihit cluster, containing:
\describe{
\item{trial}{Trial identifier.}
\item{subject}{Subject identifier.}
\item{sample}{Sample identifier.}
\item{clusterID}{Cluster identifier.}
\item{repeat_class}{The most abundant repeat class in the cluster.}
\item{repeat_class_count}{Count of the most abundant repeat class.}
\item{total_repeat_detections}{Total repeat detections in the cluster.}
\item{percentage_of_repeat_detections}{Percentage of the most abundant repeat class in the cluster.}
}
}
\description{
This is a wrapper function that goes from a raw AAVengeR multihits table and corresponding repeat table to the most representative repeat class per multihit cluster.
}
\details{
The function performs the following steps:
\enumerate{
\item Finds overlaps between \code{aavenger_mhc} and \code{repeat_table_grange} for multihit clusters using \code{find_overlaps_in_repeats()}.
\item Counts occurrences of each repeat class within each cluster using \code{count_repeat_class_per_cluster()}.
\item Extracts the most abundant repeat class for each cluster with \code{extract_most_abundant_repeat_class_per_cluster()}.
}
}
