% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_integration_percentages.R
\name{plot_integration_percentages}
\alias{plot_integration_percentages}
\title{Plot integration percentages from simulation results and experimental probabilities}
\usage{
plot_integration_percentages(
  simulation_results,
  experimental_probabilities,
  bindwidth_value,
  use_basename = TRUE
)
}
\arguments{
\item{simulation_results}{A \code{data.frame} object containing simulation results, including overlap window size, feature track file,
and simulated integration percentages. Generated by \code{rnorm_simulate_many_counts_in_features()} or \code{simulate_many_counts_in_features()}}

\item{experimental_probabilities}{A \code{data.frame} object containing experimental integration probabilities, including overlap window size,
feature track file, heatmap group, and experimental integration percentages. Generated by \code{experimental_counts_in_features()}.}

\item{bindwidth_value}{The binwidth for the histogram.}

\item{use_basename}{A logical value indicating whether to use the base name of the feature track file for plot titles (for cleaner titles).}
}
\value{
A list of ggplot objects representing the generated plots for each feature track.
}
\description{
This function generates and displays histograms of integration percentages from simulation results,
overlaid with rectangles representing experimental integration probabilities.
}
\examples{
# Example usage:
plot_integration_percentages(simulation_results, experimental_probabilities, 2, TRUE)

}
