% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabel_repeat_categories.R
\name{relabel_repeat_categories}
\alias{relabel_repeat_categories}
\title{Relabel Repeat Categories}
\usage{
relabel_repeat_categories(df)
}
\arguments{
\item{df}{A data frame containing a column \code{repeat_class}, which categorizes different types of repeats.}
}
\value{
A data frame with an additional column \code{repeat_simplified}, where repeat categories are relabeled into standardized classifications:
\describe{
\item{LINE}{Long Interspersed Nuclear Element (LINE) repeats.}
\item{SINE}{Short Interspersed Nuclear Element (SINE) repeats.}
\item{LTR}{Long Terminal Repeat (LTR) elements.}
\item{satellite}{Satellite repeats.}
\item{retrotransposon}{Retrotransposon repeats.}
\item{DNA}{DNA transposons.}
\item{unknown}{Unknown repeats.}
\item{RNA}{RNA-related repeats, including rRNA, snRNA, scRNA, tRNA, and srpRNA.}
\item{low_complexity}{Low complexity regions.}
\item{other}{Other repeat types, including RC/Helitron.}
\item{simple_repeat}{Simple repeat sequences.}
\item{not_a_repeat}{Entries without a repeat classification.}
}
}
\description{
This function simplifies repeat classification labels in a data frame by categorizing various types of repeats into broader categories. The new labels are stored in a column called \code{repeat_simplified}.
}
\details{
Relabels items in the column \code{repeat_class} generated by RepeatMasker into more simplified terms stored in \code{repeat_simplified}.
}
