% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnorm_simulate_many_counts_in_features.R
\name{rnorm_simulate_many_counts_in_features}
\alias{rnorm_simulate_many_counts_in_features}
\title{Simulate integration counts/percentages per featue and overlap window size combination using normal distribution based on previous results}
\usage{
rnorm_simulate_many_counts_in_features(
  simulation_results,
  rnorm_value,
  truncated_distribution = TRUE,
  lower_bound = 0,
  upper_bound = 100
)
}
\arguments{
\item{simulation_results}{A data frame containing previous simulation results, including overlap window size, feature track file,
simulation percentage means, and simulation percentage standard deviations. Generated by \code{simulate_many_counts_in_features()}.}

\item{rnorm_value}{The number of new simulated counts to generate per feature, overlap window size combination.}

\item{truncated_distribution}{Boolean value whether to use \code{truncnorm::rtruncnorm()} instead of \code{rnorm()}. This produces a truncated normal distribution
bounded between \code{lower_bound} and \code{upper_bound} (default 0 and 100).}

\item{lower_bound}{Lower bound of truncated distribution.}

\item{upper_bound}{upper bound of truncated distribution.}
}
\value{
A data frame containing simulated integration percentages based on the normal distribution, including overlap window size, feature track file,
and simulated integration percentages.
}
\description{
This function simulates the number of counts/percentages using the normal distribution based on the mean and standard deviation of values obtained from
\code{simulate_many_counts_in_features()}.
}
\examples{
# Example usage:
simulation_results <- data.frame(
  'overlap_window_size' = rep(c(0, 10, 1000), 2),
  'feature_track' = c(rep('feature1', 3), rep('feature2', 3)),
  'sim_n' = rep(seq(1, 3), 2),
  'count_in_feature' = c(10, 12, 8, 100, 105, 99),
  'total_sites' = 1000
) \%>\%
  dplyr::mutate(
    percentage_integrations_in_feature = 100 * (count_in_feature/total_sites)
  )
rnorm_simulate_many_counts_in_features(simulation_results, 100)

}
