% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_to_heatmap.R
\name{roc_to_heatmap}
\alias{roc_to_heatmap}
\title{ROC to Heatmap}
\usage{
roc_to_heatmap(hot_roc_result)
}
\arguments{
\item{hot_roc_result}{The multidata object produced by \code{hotROCs::ROC.ORC()}.}
}
\value{
A \code{ggplot2()} plot object.
}
\description{
Converts the output from \code{hotroc_compare_insertion_to_match()} into a \code{ggplot2()} heatmap.
}
\details{
Uses \code{format_hot_roc_result()} prior to creating ggplot object.
}
\examples{
format_hot_roc_result(hot_roc_result)

}
