% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_genome.R
\name{sample_genome}
\alias{sample_genome}
\title{Sample Genome}
\usage{
sample_genome(
  number_of_random_positions,
  chromosome_lengths,
  random_seed_value
)
}
\arguments{
\item{number_of_random_positions}{A \code{numeric} value that specifies how many different random positions will be generated.}

\item{chromosome_lengths}{A \code{data.frame} object. See \code{fasta_to_chromosome_lengths()}.}

\item{random_seed_value}{A \code{numeric} value to use as a seed for the random number generator function \code{r.seed()}.}
}
\value{
A \code{data.frame} object with two columns: \code{seqname} and \code{random_positions}.
}
\description{
Randomly sample positions in a genome.
}
\details{
Sampling is performed with replacement. The \code{data.frame} \code{chromosome_lengths} object is used so that users know what is being sampled is a dataframe of chromosomal lengths.
}
\examples{
sample_genome(number_of_random_positions, chromosome_lengths, random_seed_value)

}
