% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_gene_features.R
\name{separate_gene_features}
\alias{separate_gene_features}
\title{Separate gene features}
\usage{
separate_gene_features(gene_entry)
}
\arguments{
\item{gene_entry}{A data frame row representing a single gene entry.}
}
\value{
A data frame row containing separated gene features with columns start, end, feature, and name
}
\description{
This function separates gene features such as exons, introns, and UTRs from a gene entry.
}
\details{
Gene entry must have the following columns:
strand genomic strand
chrom chromosome name
exonStarts comma delimited list of exon positions and end with a comma (character)
exonEnds comma delimited list of exon positions and end with a comma (character)
exonCount number of exons in exonStarts/exonEnds (numeric)
cdsStartStat either cmpl or none (character)
txStart transcription start (numeric)
txEnd transcription end (numeric)
cdsStart coding start (numeric)
cdsEnd coding end (numeric)
name identifier

All numeric positions are being read in 5'->3'

Returns a dataframe with each row being a range for the category of UTR_left, UTR_right, exon, and intron
}
