% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_posid_into_chromosome_position.R
\name{split_posid_into_chromosome_position}
\alias{split_posid_into_chromosome_position}
\title{Split \code{posid} into Chromosome, Position, and Strand}
\usage{
split_posid_into_chromosome_position(df)
}
\arguments{
\item{df}{A data frame containing a \code{posid} column, where each value is a string encoding chromosome, position, and optional strand information.}
}
\value{
A data frame with new columns:
\describe{
\item{chromosome}{The chromosome component of \code{posid}.}
\item{position}{The numeric position component of \code{posid}.}
\item{extra}{A numeric extra component derived from \code{position}.}
\item{strand}{The strand information, extracted as \code{+} or \code{-} from \code{posid}.}
}
}
\description{
This function separates the \code{posid} column in a data frame into \code{chromosome}, \code{position}, and \code{extra} columns. It also extracts the strand information from \code{posid}.
}
\details{
Splits the posid column from the AAVengeR output into \code{chromosome}, \code{position}, and \code{extra} columns. The \code{posid} column will remain.
}
