% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_for_overlaps.R
\name{test_for_overlaps}
\alias{test_for_overlaps}
\title{Test for Overlaps}
\usage{
test_for_overlaps(
  matched_aavenger_sites_df,
  list_of_feature_files,
  overlap_ranges_to_test
)
}
\arguments{
\item{matched_aavenger_sites_df}{A \code{data.frame} object containing specifically formatted match and insertion sites for each heatmap_group. See \code{format_aavenger_sites()}.}

\item{list_of_feature_files}{A \code{list} object of \code{rds} or \code{RData} files containing a \code{GRanges} object called \code{epigenData}. The file path must be a prefix and the file type must be the suffix.
The data required in the \code{GRanges} object is listed below:
\describe{
\item{seqname}{name of the chromosome. Must match those in \code{matched_aavenger_sites_df}}.
\item{ranges}{An \code{IRanges} object with a genomic range.}
\item{Rle}{The strand the range is located in. Should be set to \code{*}.}}}

\item{overlap_ranges_to_test}{A \code{numeric} \code{vector} of the different genomic windows to use in the overlap test. Values must be integers.}
}
\value{
A dataframe containing the original \code{matched_aavenger_sites_df} and a column of overlap counts per each file from \code{list_of_feature_files}.
}
\description{
Count the number of times an integration site overlaps with a genomic feature.
}
\examples{
test_for_overlaps(matched_aavenger_sites_df, list_of_feature_files = list_of_feature_filenames,data_dir=file.path('/path/to/feature_filenames'),overlap_ranges_to_test = c(1000,10000,1000000))

}
