% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_mhc_table.R
\name{unnest_mhc_table}
\alias{unnest_mhc_table}
\title{Unnest MHC Table}
\usage{
unnest_mhc_table(aavenger_mhc)
}
\arguments{
\item{aavenger_mhc}{A data frame representing the AAvengeR MHC multihits cluster table, which must contain a \code{posids} column of class \code{list}.}
}
\value{
A data frame with the \code{posids} column unnested and renamed to \code{posid}.
}
\description{
This function takes an AAvengeR MHC table with a \code{posids} column as a list and unnests it. The \code{posids} column is renamed to \code{posid}.
}
\details{
This function takes an AAvengeR MHC table with a \code{posids} column as a list and unnests it. The \code{posids} column is renamed to \code{posid}.
}
