% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_overlaps.R
\name{extract_overlaps}
\alias{extract_overlaps}
\title{Extract overlapping regions between two sets of genomic ranges}
\usage{
extract_overlaps(query, subject, mark_query = TRUE, mark_subject = TRUE)
}
\arguments{
\item{query}{A GenomicRanges object representing the query set. Must have start, end, strand columns.}

\item{subject}{A GenomicRanges object representing the subject set. Must have start, end, strand columns.}

\item{mark_query}{Logical indicating whether to mark columns of the query set with "_query" suffix.}

\item{mark_subject}{Logical indicating whether to mark columns of the subject set with "_subject" suffix.}
}
\value{
A data frame containing all overlapping regions between the query and subject sets.
}
\description{
This function identifies overlapping regions between a query set and a subject set of genomic ranges.
}
\details{
Output is a data table with all instances of query in subject, even if query has multiple hits in subject.
All columns from query and subject are joined in the output via cbind.
}
